feat: docker and docker-compose

+++ tg_bot/modules/sql/afk_sql.py
@@ -1,6 +1,6 @@
 import threading
 
-from sqlalchemy import Column, UnicodeText, Boolean, Integer
+from sqlalchemy import Column, UnicodeText, Boolean, Integer, BigInteger
 
 from tg_bot.modules.sql import BASE, SESSION
 
@@ -8,7 +8,7 @@ from tg_bot.modules.sql import BASE, SESSION
 class AFK(BASE):
     __tablename__ = "afk_users"
 
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     is_afk = Column(Boolean)
     reason = Column(UnicodeText)
